+++ lib/checkAlwaysNever.m
@@ -35,8 +35,8 @@ function [endUser, applianceList] = checkAlwaysNever(endUser, applianceList, cou
 			% Assign usage vector, duc (dailyUsageCount), and wuc(weeklyUsageCount)
 			endUser.appliances.(string(endUser.properties.alwaysAppliances(index))).usageArray =...
 																																							repmat(generateUsageVector(),countDays,1);
-			endUser.appliances.(string(endUser.properties.alwaysAppliances(index))).duc = 0;
-			endUser.appliances.(string(endUser.properties.alwaysAppliances(index))).wuc = 0;
+			endUser.appliances.(string(endUser.properties.alwaysAppliances(index))).duc = uint16(0);
+			endUser.appliances.(string(endUser.properties.alwaysAppliances(index))).wuc = uint16(0);
 		end
 	end
 end
